Split preemption integration tests to mitigate the timeout
